requirements.txt
faiss-cpu
scikit-learn
numpy
